+++ packages/app/src/components/catalog/EntityPage.tsx
@@ -53,7 +53,7 @@ import {
 
 import { TechDocsAddons } from '@backstage/plugin-techdocs-react';
 import { ReportIssue } from '@backstage/plugin-techdocs-module-addons-contrib';
-import { DaytonaOverviewContent } from '@adityasinghal26/backstage-plugin-daytona';
+import { DaytonaOverviewContent } from '@daytonaio/backstage-plugin-daytona';
 
 const techdocsContent = (
   <EntityTechdocsContent>